Refactor - split the data model definitions into separate model.py
